query notes

+++ src/models/compras.js
@@ -7,7 +7,7 @@ module.exports = (sequelize) => {
     pedido: {type: DataTypes.ARRAY(DataTypes.JSON)},
     total: {type: DataTypes.STRING},
     direccionDeEnvio: {type: DataTypes.STRING},
-    celular: {type: DataTypes.INTEGER},
+    celular: {type: DataTypes.BIGINT},
     id: {type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true}
   }, {timestamps: false })
 };
